fix issue with root pointer inputs  causing panics
If the root type (the type T provided to goscanql.RowsToStructs)
was of type pointer, it would cause a panic because the fields.obj
type was expected to be a pointer to the root type (i.e. *User)
instead of the pointer to any other number of pointers to the root
type (e.g. **User).